+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to
 
 + Added fallback to `application/octet-stream` for unknown MIME types
 
+### Fixed
+
++ Fixed typo in readme
+
 ## [2.0.0] - 2019-08-25
 
 ### Added