+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [2.1.0] - 2019-09-07
+
 ### Added
 
 + Added fallback to `application/octet-stream` for unknown MIME types
@@ -68,6 +70,7 @@ and this project adheres to
 
 + Initial release
 
-[Unreleased]: https://github.com/mserajnik/hyve/compare/2.0.0...develop
+[Unreleased]: https://github.com/mserajnik/hyve/compare/2.1.0...develop
+[2.1.0]: https://github.com/mserajnik/hyve/compare/2.0.0...2.1.0
 [2.0.0]: https://github.com/mserajnik/hyve/compare/1.1.0...2.0.0
 [1.1.0]: https://github.com/mserajnik/hyve/compare/1.0.0...1.1.0